create separate classes for functions
